requirements.txt
Flask==3.1.2
psycopg==3.2.12
python-dotenv==1.2.1